LPS-174758 LPS-200418 portal-search: Apply new way to use and move to new package
