+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/suggestions/spi/asah/individuals/AsahRecentSitesIndividualsContributor.java
@@ -3,23 +3,19 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.suggestions.spi;
+package com.liferay.portal.search.internal.suggestions.spi.asah.individuals;
 
-import com.liferay.analytics.settings.rest.manager.AnalyticsSettingsManager;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.search.internal.spi.model.index.contributor.GroupUtil;
 import com.liferay.portal.search.rest.dto.v1_0.SuggestionsContributorConfiguration;
 import com.liferay.portal.search.spi.suggestions.SuggestionsContributor;
-import com.liferay.portal.search.suggestions.SuggestionBuilderFactory;
 import com.liferay.portal.search.suggestions.SuggestionsContributorResults;
-import com.liferay.portal.search.suggestions.SuggestionsContributorResultsBuilderFactory;
+import com.liferay.portal.search.suggestions.spi.constants.AsahSuggestionsConstants;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Gustavo Lima
@@ -42,24 +38,28 @@ public class AsahRecentSitesIndividualsContributor
 			suggestionsContributorConfiguration) {
 
 		return getSuggestionsContributorResults(
-			_analyticsSettingsManager, "recent-sites", "individuals",
-			searchContext, "visits,lastVisitDate,firstVisitDate,groupId",
-			_suggestionBuilderFactory, suggestionsContributorConfiguration,
-			_suggestionsContributorResultsBuilderFactory,
-			_portalUtil.getUserId(liferayPortletRequest));
+			StringBundler.concat(
+				AsahSuggestionsConstants.INDIVIDUALS, "/",
+				getHashedEmail(portal.getUserId(liferayPortletRequest))),
+			AsahSuggestionsConstants.RECENT_SITES, searchContext,
+			"visits,lastVisitDate,firstVisitDate,groupId",
+			suggestionsContributorConfiguration);
 	}
 
-	@Reference
-	private AnalyticsSettingsManager _analyticsSettingsManager;
+	@Override
+	protected String getAssetURL(
+		String destinationBaseURL, JSONObject itemJSONObject) {
 
-	@Reference
-	private PortalUtil _portalUtil;
+		return itemJSONObject.getString("url");
+	}
 
-	@Reference
-	private SuggestionBuilderFactory _suggestionBuilderFactory;
+	@Override
+	protected String getText(
+		String destinationBaseURL, JSONObject itemJSONObject) {
 
-	@Reference
-	private SuggestionsContributorResultsBuilderFactory
-		_suggestionsContributorResultsBuilderFactory;
+		// api not return the site name
+
+		return itemJSONObject.getString("site name");
+	}
 
 }
